Generalize exported WriteStream vs WriteStdin
